Revert "[WIP] Merging this repo with the greenland master branch"
