+++ .gitignore
@@ -35,7 +35,7 @@
 *.i*86
 *.x86_64
 *.hex
-greenland*
+greenland
 tests
 
 # Debug files
