+++ Makefile
@@ -2,15 +2,14 @@ CC=gcc
 CFLAGS=-Wall -Wextra -Wformat-nonliteral -Wcast-align -Wpointer-arith \
 -Wbad-function-cast -Wmissing-prototypes -Wstrict-prototypes \
 -Wmissing-declarations -Winline -Wundef -Wnested-externs -Wcast-qual -Wshadow \
--Wwrite-strings -Wno-unused-parameter -Wno-declaration-after-statement \
--Wfloat-equal -pedantic -ansi -Werror -std=c17
+-Wwrite-strings -Wno-unused-parameter -Wfloat-equal -pedantic -ansi -Werror
 SRCDIR=src
 LIBDIR=lib
 TESTDIR=test
 DOCDIR=doc
-BIN=greenland_*
+BIN=greenland
 
-make: test client server
+make: test main
 	doxygen Doxyfile
 	astyle --project $(LIBDIR)/*.c $(LIBDIR)/*.h $(SRCDIR)/*.c $(TESTDIR)/*.c
 	./tests
@@ -21,17 +20,11 @@ test: tree.o menuoption.o test.o
 test.o: $(TESTDIR)/unit_tests.c
 	$(CC) $(CFLAGS) -c $(TESTDIR)/*.c
 
-client: tree.o menuoption.o client.o
-	$(CC) $(CFLAGS) client.o menuoption.o tree.o tree_state.o -o greenland_client
+main: tree.o menuoption.o main.o
+	$(CC) $(CFLAGS) main.o menuoption.o tree.o tree_state.o -o greenland 
 
-client.o: $(SRCDIR)/client.c
-	$(CC) $(CFLAGS) -c $(SRCDIR)/client.c
-
-server: tree.o server.o
-	$(CC) $(CFLAGS) server.o tree.o tree_state.o -o greenland_server
-
-server.o: $(SRCDIR)/server.c
-	$(CC) $(CFLAGS) -c $(SRCDIR)/server.c
+main.o: $(SRCDIR)/main.c
+	$(CC) $(CFLAGS) -c $(SRCDIR)/main.c
 
 tree.o: tree_state.o $(LIBDIR)/tree.h $(LIBDIR)/tree.c
 	$(CC) $(CFLAGS) -c $(LIBDIR)/tree.h $(LIBDIR)/tree.c
